--- packages/vite-glob-routes/src/internal.ts
@@ -1,2 +1,0 @@
-export * from "./react-router-internal";
-export * from "./hattip-internal";
