+++ packages/vite-glob-routes/src/internal/hattip.ts
