+++ packages/vite-glob-routes/src/internal/index.ts
@@ -1,0 +1,2 @@
+export * from "./react-router";
+export * from "./hattip";
