+++ packages/vite-glob-routes/src/internal/react-router.ts
