+++ packages/vite-glob-routes/src/internal/utils.ts
