+++ packages/vite-glob-routes/src/react-router.ts
@@ -1,4 +1,4 @@
-import type { createGlobPageRoutes } from "./react-router-internal";
+import type { createGlobPageRoutes } from "./internal/react-router";
 
 export const globPageRoutes = (() => {
   throw new Error(
