+++ packages/vite-glob-routes/tsup.config.ts
@@ -3,9 +3,9 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: [
     "src/index.ts",
-    "src/internal.ts",
     "src/react-router.ts",
     "src/hattip.ts",
+    "src/internal/index.ts",
   ],
   format: ["esm", "cjs"],
   dts: true,
